refactor(migration): type legacy_references foreign keys as BigInteger
The legacy_references primary key is BigInteger Identity, so foreign keys
that reference it must match: cloned_from_id and the child tables'
reference_id. Mirrors the legacy_samples convention. Outbound FKs to
Integer serial PKs (user_id, upload_id, task_id, group_id) stay Integer.

+++ assets/alembic/versions/a73a2668403f_create_references_table.py
@@ -31,7 +31,7 @@ def upgrade() -> None:
         sa.Column("source_types", JSONB(), nullable=False),
         sa.Column("user_id", sa.Integer(), nullable=True),
         sa.Column("upload_id", sa.Integer(), nullable=True),
-        sa.Column("cloned_from_id", sa.Integer(), nullable=True),
+        sa.Column("cloned_from_id", sa.BigInteger(), nullable=True),
         sa.Column("task_id", sa.Integer(), nullable=True),
         sa.PrimaryKeyConstraint("id"),
         sa.UniqueConstraint("legacy_id"),
@@ -42,7 +42,7 @@ def upgrade() -> None:
     )
     op.create_table(
         "legacy_reference_groups",
-        sa.Column("reference_id", sa.Integer(), nullable=False),
+        sa.Column("reference_id", sa.BigInteger(), nullable=False),
         sa.Column("group_id", sa.Integer(), nullable=False),
         sa.Column("build", sa.Boolean(), nullable=False),
         sa.Column("modify", sa.Boolean(), nullable=False),
@@ -55,7 +55,7 @@ def upgrade() -> None:
     )
     op.create_table(
         "legacy_reference_users",
-        sa.Column("reference_id", sa.Integer(), nullable=False),
+        sa.Column("reference_id", sa.BigInteger(), nullable=False),
         sa.Column("user_id", sa.Integer(), nullable=False),
         sa.Column("build", sa.Boolean(), nullable=False),
         sa.Column("modify", sa.Boolean(), nullable=False),
